--- KvasovField.py
@@ -1,16 +1,0 @@
-weight=0
-height=0
-size=[weight,height]
-teleport1_pos=[0,0]
-teleport2_pos=[2,2]
-teleport_pos=[teleport1_pos,teleport2_pos]
-
-def draw(size):
-    #отрисовка
-    raise NotImplemented
-def collision():
-    #проверка на возможность передвижения
-    raise NotImplemented
-def get_teleport_pos(teleport_pos):
-    #выводит координаты телепортов
-    raise NotImplemented
