--- KvasovSeed.py
@@ -1,16 +1,0 @@
-grains = [[1, 1], [0, 0]]  # матрица зерен(чтобы не перепутать с seed другое название(хз нужно ли это))
-count = 0
-
-
-def draw():
-    # отрисовка всех зерен
-    raise NotImplemented
-
-
-def delete_seed():
-    # удалаем зерно
-    raise NotImplemented
-
-
-
-
