+++ field.py
@@ -1,18 +1,18 @@
-class Field:
-    def __init__(self):
-        self.weight = 0
-        self.height = 0
-        self.size = [self.weight, self.height]
-        self.teleport1_pos = [0, 0]
-        self.teleport2_pos = [1, 1]
-
-    def draw(self, size):
-        pass
-
-    def collision(self):
-        # проверка на возможность передвижения
-        pass
-
-    def get_teleport_pos(self):
-        # выводит координаты телепортов
-        pass
+weight=0
+height=0
+size=[weight,height]
+teleport1_pos=[0,0]
+teleport2_pos=[2,2]
+teleport_pos=[teleport1_pos,teleport2_pos]
+def get_plan():
+    #считывание JSON-а файла
+    raise NotImplemented
+def draw(size):
+    #отрисовка
+    raise NotImplemented
+def check_move():
+    #проверка на возможность передвижения
+    raise NotImplemented
+def get_teleport_pos(teleport_pos):
+    #выводит координаты телепортов
+    raise NotImplemented
