--- main.py
@@ -1,11 +1,0 @@
-import pygame
-import game
-
-
-def main():
-    pygame.init()
-    game.start()
-
-
-if __name__ == "__main__":
-    main()
