+++ pacman.py
@@ -1,44 +1,34 @@
-import pacman
-
-pos = (0, 0)
-direction = (1, 0)
-image = None #удет добавлено позже
+posx = 0
+posy = 0
+pos = [posx, posy]
+image = None
 score = 0
+direction = (1, 0)
 lives = 3
 
 
-
 def move():
-    raise NotImplementedError      #  reqest to field (dir, pos) - next_pos
+    raise NotImplemented
 
-
-def check_ghost(self):
-    # роверка на привидение
-    # lifes -here
-    #lifes - 0 games.run = false
-    raise NotImplementedError
-
-
-def check_seed():
-    raise NotImplementedError
 
 def step():
     move()
     check_ghost()
     check_seed()
 
-def draw():
+
+def check_ghost():
+    raise NotImplemented
+
+
+def change_direction():
+    raise NotImplemented
+
+
+def draw(image):
     # Отрисовка пакмана.
-    pass
-
-def change_direction(button):
-    global pos
-    pos = (9, 0)
-    # считывает нажатие кнопки и меняет направление пакмана
-    pass
+    raise NotImplemented
 
 
-
-
-
-
+def check_seed():
+    raise NotImplemented
